Hackathon 2020: Move RHS assembly of adjoint equation to separate assembler. Add example parameterfile to benchmarks folder.
