+++ benchmarks/adjoint/adjointkernel_degree4.prm
@@ -1,0 +1,145 @@
+# A simple setup for convection in a quarter of a 2d shell. See the
+# manual for more information.
+
+
+set Dimension                              = 2
+set Use years in output instead of seconds = true
+set End time                               = 0 #1.5e9
+set Output directory                       = output_4
+set Pressure normalization                = surface
+set Surface pressure                      = 0
+set Nonlinear solver scheme = Stokes adjoint
+
+subsection Adjoint problem
+  set Read points from file             = false #true
+  set Number of iterations in adjoint inversion  =  0
+  set Factor to update the density properties    =  1
+  set Factor to update the viscosity properties  =  1
+end
+
+
+
+subsection Material model
+  # additive means after each iteration we add the value stored in the 
+  # composition field to the existing density / multiply the viscosity 
+  # with the new composition field
+  set Model name = additive
+
+  subsection Additive model
+#    set Base model = simple
+     set Base model additive = simple
+  end
+
+  subsection Simple model
+    set Thermal expansion coefficient = 4e-5
+    set Viscosity                     = 1e21
+    set Thermal viscosity exponent    = 3
+    set Reference temperature         = 1600
+  end
+
+end
+
+# to run sensibly with stokes adjoint the following things stay:
+# Composition polynomial degree
+# Additive model
+# 2 compositional fields with exactly those names 
+subsection Discretization
+  set Composition polynomial degree        = 0
+  set Use discontinuous composition discretization = true
+end
+
+subsection Compositional fields
+  set Number of fields = 2
+  set Names of fields = density_term, viscosity_factor
+end
+
+subsection Initial composition model
+  set Model name = function
+
+  subsection Function
+    # before the iteration the density term should be 0 and the
+    # viscosity prefactor should be 1
+    # this means the addition to density at the first iteration is 0 
+    # while the multiplication to viscosity is 1
+    # these scalar values are prescribed everywhere in 3D field
+    # during iteration the composition value will reflect the incremental addition
+    # multiplication
+    set Function expression = 0 ; 1
+  end
+end
+
+
+
+subsection Geometry model
+  set Model name = spherical shell
+
+  subsection Spherical shell
+    set Inner radius  = 3481000
+    set Outer radius  = 6336000
+  end
+end
+
+
+subsection Boundary velocity model
+    set Zero velocity boundary indicators       =
+  set Tangential velocity boundary indicators = inner, outer
+  set Prescribed velocity boundary indicators =
+end
+
+subsection Boundary temperature model
+  set Fixed temperature boundary indicators   = inner, outer
+end
+
+
+subsection Nullspace removal
+  set Remove nullspace = net rotation
+end
+
+subsection Boundary temperature model
+  set Model name = spherical constant
+  subsection Spherical constant
+    set Inner temperature = 4273
+    set Outer temperature = 973
+  end
+end
+
+
+subsection Initial temperature model
+  set Model name = spherical hexagonal perturbation
+
+  subsection Spherical hexagonal perturbation
+    set Angular mode = 4
+  end
+end
+
+
+subsection Gravity model
+  set Model name = radial constant
+
+  subsection Radial constant
+    set Magnitude = 10
+  end
+
+end
+
+
+subsection Mesh refinement
+  set Initial global refinement          = 5
+  set Initial adaptive refinement        = 0
+  set Strategy                           = temperature
+  set Time steps between mesh refinement = 15
+end
+
+
+subsection Postprocess
+  set List of postprocessors = adjoint kernels, visualization, dynamic topography #, adjoint kernels, velocity statistics, temperature statistics, heat flux statistics, depth average
+
+#  subsection Visualization
+#    set Output format                 = vtu
+#    set List of output variables      = dynamic topography
+#    set Time between graphical output = 1e6
+#    set Number of grouped files       = 0
+#    set Interpolate output            = false
+#  end
+
+end
