+++ include/aspect/simulator/assemblers/stokes.h
@@ -217,19 +217,6 @@ namespace aspect
         execute(internal::Assembly::Scratch::ScratchBase<dim>   &scratch_base,
                 internal::Assembly::CopyData::CopyDataBase<dim> &data_base) const override;
     };
-
-    /**
-     */
-    template <int dim>
-    class StokesAdjointRHS : public Assemblers::Interface<dim>,
-      public SimulatorAccess<dim>
-    {
-      public:
-        virtual
-        void
-        execute(internal::Assembly::Scratch::ScratchBase<dim>   &scratch,
-                internal::Assembly::CopyData::CopyDataBase<dim> &data) const;
-    };
   }
 }
