--- source/postprocess/visualization/adjoint_velocity.cc
@@ -1,84 +1,0 @@
-/*
-  Copyright (C) 2011 - 2015 by the authors of the ASPECT code.
-
-  This file is part of ASPECT.
-
-  ASPECT is free software; you can redistribute it and/or modify
-  it under the terms of the GNU General Public License as published by
-  the Free Software Foundation; either version 2, or (at your option)
-  any later version.
-
-  ASPECT is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU General Public License for more details.
-
-  You should have received a copy of the GNU General Public License
-  along with ASPECT; see the file doc/COPYING.  If not see
-  <http://www.gnu.org/licenses/>.
-*/
-
-
-#include <aspect/postprocess/visualization/adjoint_velocity.h>
-//#include <aspect/simulator.h>
-//#include <aspect/simulator_access.h>
-
-
-namespace aspect
-{
-  namespace Postprocess
-  {
-    namespace VisualizationPostprocessors
-    {
-      template <int dim>
-      AdjointVelocity<dim>::
-      AdjointVelocity ()
-        :
-        DataPostprocessorVector<dim> ("adjoint_velocity",
-                                      update_q_points)
-      {}
-
-
-
-      template <int dim>
-      void
-      AdjointVelocity<dim>::
-      compute_derived_quantities_vector (const std::vector<Vector<double> >              &solution_values,
-                                         const std::vector<std::vector<Tensor<1,dim> > > &,
-                                         const std::vector<std::vector<Tensor<2,dim> > > &,
-                                         const std::vector<Point<dim> > &,
-                                         const std::vector<Point<dim> >                  &evaluation_points,
-                                         std::vector<Vector<double> >                    &computed_quantities) const
-      {
-        const unsigned int n_quadrature_points = solution_values.size();
-        Assert (computed_quantities.size() == n_quadrature_points,    ExcInternalError());
-        Assert (computed_quantities[0].size() == dim,                   ExcInternalError());
-        Assert (solution_values[0].size() == this->introspection().n_components,           ExcInternalError());
-
-        for (unsigned int q=0; q<n_quadrature_points; ++q)
-          {
-            for (unsigned int d = 0; d < dim; ++d)
-              {
-                computed_quantities[q](d) = solution_values[q][this->introspection().component_indices.velocities[d]];;
-              }
-          }
-
-      }
-    }
-  }
-}
-
-
-// explicit instantiations
-namespace aspect
-{
-  namespace Postprocess
-  {
-    namespace VisualizationPostprocessors
-    {
-      ASPECT_REGISTER_VISUALIZATION_POSTPROCESSOR(AdjointVelocity,
-                                                  "adjoint velocity",
-                                                  "A visualization output object that outputs the gravity vector.")
-    }
-  }
-}
